added card deck editing, added new mob, added new city buildings, added tower info panels, new enemies, new towers,

+++ Assets/Resources/Codebase/Enemies/HealthModules/AbstractDamageRecievingModule.cs
@@ -18,5 +18,5 @@ public abstract class AbstractDamageRecievingModule
     /// </summary>
     /// <param name="damage">flat incoming damage</param>
     /// <returns>calculated damage</returns>
-    public abstract int CalculateRecievedDamage(int damage);
+    public abstract float CalculateRecievedDamage(float damage);
 }
